# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestão de Vendas",
        "version": "1.0.0"
    },
    "paths": {
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Lista clientes",
                "operationId": "de55d50d1b708d9e5c63ee3f8fe8f54a",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cria um cliente",
                "operationId": "4ac6cdfa8d6c188fa2cb4f3263d4318a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Jonathan Bayer"
                                    },
                                    "cpf_cnpj": {
                                        "type": "string",
                                        "example": "30455078447"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "(575) 535-4495"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "thiago@example.com"
                                    },
                                    "endereco": {
                                        "type": "string",
                                        "example": "539 Senger Port"
                                    },
                                    "cidade": {
                                        "type": "string",
                                        "example": "Port Edytheburgh"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "IN"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "79231"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Mostra um cliente",
                "operationId": "0d3826d79390929601421b5810b61231",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualiza um cliente",
                "operationId": "74956b09f3ea120b5dd531c73776ca11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Jonathan Bayer"
                                    },
                                    "cpf_cnpj": {
                                        "type": "string",
                                        "example": "30455078447"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "(575) 535-4495"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "thiago@example.com"
                                    },
                                    "endereco": {
                                        "type": "string",
                                        "example": "539 Senger Port"
                                    },
                                    "cidade": {
                                        "type": "string",
                                        "example": "Port Edytheburgh"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "IN"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "79231"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Exclui um cliente",
                "operationId": "d15e304c609752c9f409ab81072fb52e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Excluído"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/financeiro/lancamentos": {
            "get": {
                "tags": [
                    "Financeiro"
                ],
                "summary": "Lista lançamentos financeiros",
                "operationId": "a75a7e05447804684ee6b8048fccad74",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por status: aberto, pago ou cancelado",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "aberto",
                                "pago",
                                "cancelado"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de lançamentos"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/financeiro/{id}/pagar": {
            "put": {
                "tags": [
                    "Financeiro"
                ],
                "summary": "Marca um lançamento como pago",
                "operationId": "e936ea7e16a5458c6c8e0d337ee544b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lançamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lançamento pago com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Lançamento não encontrado"
                    },
                    "500": {
                        "description": "Erro ao processar pagamento"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}